Added a metiatR query. Added the bets dbset. added migrations

+++ OnlineAuction.API/Models/Auth/RoleClaim.cs
@@ -2,7 +2,7 @@
 
 namespace OnlineAuction.Domain.Auth
 {
-    public class RoleClaim : IdentityRoleClaim<int>
+    public class RoleClaim : IdentityRoleClaim<long>
     {
     }
 }
