+++ OnlineAuction.API/Models/Auth/User.cs
@@ -3,9 +3,10 @@ using System.Globalization;
 
 namespace OnlineAuction.Domain.Auth
 {
-    public class User : IdentityUser<int>
+    public class User : IdentityUser<long>
     {
         public int AccountBalance { get; set; }
+
 
     }
 }
